git-svn-id: http://svn.openzap.org/svn/openzap/branches/sangoma_boost@939 a93c3328-9c30-0410-af19-c9cd2b2d52af

+++ libs/freetdm/src/ozmod/ozmod_isdn/ozmod_isdn.c
@@ -992,7 +992,7 @@ static L3INT zap_isdn_931_34(void *pvt, L2UCHAR *msg, L2INT mlen)
 							}
 
 							/* TODO: make this more safe with strncat() */
-							pos = strlen(zchan->caller_data.dnis.digits);
+							pos = (int)strlen(zchan->caller_data.dnis.digits);
 							strcat(&zchan->caller_data.dnis.digits[pos],    (char *)callednum->Digit);
 
 							/* update timer */
